Added unit tests for get_file_hash function
